requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
opencv-python>=4.9.0.80
mss>=9.0.1
pytesseract>=0.3.10
PySide6>=6.8.1
websockets>=12.0
google-genai>=0.5.0
pydantic>=2.6.3
python-dotenv>=1.0.1